+++ src/libs/blueprint/conduit_blueprint_mesh.cpp
@@ -2670,17 +2670,16 @@ mesh::generate_index(const Node &mesh,
             idx_matset["topology"] = matset["topology"].as_string();
 
             // support different flavors of valid matset protos
+            //
+            // if we have material_map (node with names to ids)
+            // use it in the index
             if(matset.has_child("material_map"))
             {
-                NodeConstIterator mats_itr = matset["material_map"].children();
-                while(mats_itr.has_next())
-                {
-                    mats_itr.next();
-                    idx_matset["materials"][mats_itr.name()];
-                }
+                idx_matset["material_map"] = matset["material_map"];
             }
             else if(matset.has_child("materials"))
             {
+                // NOTE: I believe path is deprecated ... 
                 NodeConstIterator mats_itr = matset["materials"].children();
                 while(mats_itr.has_next())
                 {
@@ -2690,11 +2689,14 @@ mesh::generate_index(const Node &mesh,
             }
             else if(matset.has_child("volume_fractions"))
             {
+                // we don't have material_map (node with names to ids)
+                // so mapping is implied from node order, construct
+                // an actual map that follows the implicit order
                 NodeConstIterator mats_itr = matset["volume_fractions"].children();
                 while(mats_itr.has_next())
                 {
                     mats_itr.next();
-                    idx_matset["materials"][mats_itr.name()];
+                    idx_matset["material_map"][mats_itr.name()] = mats_itr.index();
                 }
             }
             else // surprise!
@@ -4512,12 +4514,48 @@ mesh::topology::shape::verify(const Node &shape,
 //-----------------------------------------------------------------------------
 
 //-----------------------------------------------------------------------------
+// helper to verify a matset material_map
+//-----------------------------------------------------------------------------
+bool verify_matset_material_map(const std::string &protocol,
+                                const conduit::Node &matset,
+                                conduit::Node &info)
+{
+    bool res = verify_object_field(protocol, matset, info, "material_map");
+
+    if(res)
+    {
+        // we already know we have an object, children should be 
+        // integer scalars
+        NodeConstIterator itr = matset["material_map"].children();
+        while(itr.has_next())
+        {
+            const Node &curr_child = itr.next();
+            if(!curr_child.dtype().is_integer())
+            {
+                log::error(info,
+                           protocol,
+                           log::quote("material_map") +
+                           "child " +
+                           log::quote(itr.name()) +
+                           " is not an integer leaf.");
+                res = false;
+            }
+        }
+    }
+
+    log::validation(info, res);
+
+    return res;
+}
+
+//-----------------------------------------------------------------------------
 bool
 mesh::matset::verify(const Node &matset,
                      Node &info)
 {
     const std::string protocol = "mesh::matset";
     bool res = true, vfs_res = true;
+    bool mat_map_is_optional = true;
     info.reset();
 
     res &= verify_string_field(protocol, matset, info, "topology");
@@ -4535,12 +4573,11 @@ mesh::matset::verify(const Node &matset,
             verify_number_field(protocol, matset, info, "volume_fractions"))
         {
             log::info(info, protocol, "detected uni-buffer matset");
+            // materials_map is not optional in this case, signal
+            // for opt check down the line
+            mat_map_is_optional = false;
 
             vfs_res &= verify_integer_field(protocol, matset, info, "material_ids");
-            // TODO(JRC): Add a more in-depth verifier for 'material_map' that
-            // verifies that it's one level deep and that each child child houses
-            // an integer-style array.
-            vfs_res &= verify_object_field(protocol, matset, info, "material_map");
             vfs_res &= blueprint::o2mrelation::verify(matset, info);
 
             res &= vfs_res;
@@ -4571,6 +4608,47 @@ mesh::matset::verify(const Node &matset,
 
             res &= vfs_res;
             log::validation(vfs_info, vfs_res);
+        }
+    }
+
+    if(!mat_map_is_optional && !matset.has_child("material_map"))
+    {
+        log::error(info, protocol,
+            "'material_map' is missing (required for uni-buffer matsets) ");
+        res &= false;
+    }
+
+    if(matset.has_child("material_map"))
+    {
+        if(mat_map_is_optional)
+        {
+            log::optional(info, protocol, "includes material_map");
+        }
+
+        res &= verify_matset_material_map(protocol,matset,info);
+
+        // for cases where vfs are an object, we expect the material_map child 
+        // names to be a subset of the volume_fractions child names
+        if(matset.has_child("volume_fractions") &&
+           matset["volume_fractions"].dtype().is_object())
+        {
+            NodeConstIterator itr =  matset["material_map"].children();
+            while(itr.has_next())
+            {
+                itr.next();
+                std::string curr_name = itr.name();
+                if(!matset["volume_fractions"].has_child(curr_name))
+                {
+                    std::ostringstream oss;
+                    oss << "'material_map' hierarchy must be a subset of "
+                           "'volume_fractions'. " 
+                           " 'volume_fractions' is missing child '"
+                           << curr_name 
+                           <<"' which exists in 'material_map`" ;
+                    log::error(info, protocol,oss.str());
+                    res &= false;
+                }
+            }
         }
     }
 
@@ -4675,7 +4753,18 @@ mesh::matset::index::verify(const Node &matset_idx,
     // performed on the "materials" field.
 
     res &= verify_string_field(protocol, matset_idx, info, "topology");
-    res &= verify_object_field(protocol, matset_idx, info, "materials");
+
+    // 2021-1-29 cyrush:
+    // prefer new "material_map" index spec, vs old "materials"
+    if(matset_idx.has_child("material_map"))
+    {
+        res &= verify_matset_material_map(protocol,matset_idx,info);
+    }
+    else
+    {
+        res &= verify_object_field(protocol, matset_idx, info, "materials");
+    }
+
     res &= verify_string_field(protocol, matset_idx, info, "path");
 
     log::validation(info, res);